requirements.txt
# Core AI/LLM libraries
langchain>=0.1.0
openai>=1.0.0
anthropic>=0.8.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web framework (optional)
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.8.0
